Move nh5 to evm
nh5.py contains the persistent definition of everything in IC (e.g,
the pytables classes) and belongs to the evm package. It has been
moved there and the code correspondingly adapted.
